+++ public/js/tmsm-aquatonic-course-booking-public.js
@@ -553,15 +553,11 @@ var TmsmAquatonicCourseApp = TmsmAquatonicCourseApp || {};
 
   TmsmAquatonicCourseApp.SelectedDataView = Backbone.View.extend( {
     el: '#tmsm-aquatonic-course-booking-confirm-container',
-    cancelButton: '#tmsm-aquatonic-course-booking-cancel',
-    confirmButton: '#tmsm-aquatonic-course-booking-confirm',
     errorElement: '#tmsm-aquatonic-course-booking-confirm-error',
 
     initialize: function() {
       console.log('SelectedDataView initialize');
       this.hideError();
-      this.hideConfirm();
-      this.hideCancel();
       this.listenTo(this.model, 'change', this.change);
     },
 
@@ -630,7 +626,6 @@ var TmsmAquatonicCourseApp = TmsmAquatonicCourseApp || {};
         this.showConfirm();
       }
       else{
-        this.hideConfirm();
       }
 
 
@@ -653,11 +648,8 @@ var TmsmAquatonicCourseApp = TmsmAquatonicCourseApp || {};
     showConfirm: function(){
       console.log('SelectedDataView showConfirm 02');
       TmsmAquatonicCourseApp.animateTransition( $(TmsmAquatonicCourseApp.form_fields.summary_field));
-      $( this.confirmButton ).show();
     },
-    hideConfirm: function(){
-      $( this.confirmButton ).hide();
-    },
+
 
     showCancel: function(){
       console.log('SelectedDataView showCancel');